feat: add preview component to view details
