--- src/app/preview/preview.component.html
@@ -1,1 +1,0 @@
-<p>preview works!</p>
