--- src/app/preview/preview.component.ts
@@ -1,12 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-preview',
-  templateUrl: './preview.component.html',
-  styleUrls: ['./preview.component.scss'],
-})
-class PreviewComponent implements OnInit {
-  ngOnInit(): void {
-  }
-}
-export default PreviewComponent;
